# Improve the performance of network results and network serialization