This PR brings many performance improvements for results access on the network and for network serialization. The optimization is mainly based on looping over lists instead of numpy arrays which is much more performant for small sized arrays like in our case.

Iterating over a small list, of say voltages, is 4X faster than iterating over an array:

```ipy
[In]:  %%timeit
       [[v.real, v.imag] for v in voltages]
[Out]: 1.62 μs ± 32 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
vs
```ipy
[In]:  %%timeit
       [[v.real, v.imag] for v in voltages.tolist()]
[Out]: 429 ns ± 6.77 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

Similarly, computing the complex power in a list is 2X faster than using array multiplication.

This resulted in a measured speedup of 20% when accessing `en.res_buses`, `en.res_buses_voltages`, `en.res_lines`, `en.res_transformers` and `en.res_loads` in an MV feeder with 351 buses, 318 lines, 31 transformers and 498 loads. The same network showed a 15% speedup when serialized to a dict with the results.